Fixed import errors on empty password fields
Empty password (or other custom entries in bitwarden) fields get converted to the value `null`, while the Schema doesnt't allow this.
Also Bitwarden added a new type of additonal field in an entry (an alias) for selecting the correct field during autofill in a browser. This field was not handled by the script and resulted in the `Type` field beeing `null`, thus causing errors.