[ci] enable script file extraction
When the dynamic folder reindex job runs, extract the `Script` and
`DynamicCredentialScript` contents to files in this repo, to enable
easier review of the logic bundled within `.rdfe`/`.rdfx` files.

Next to each `Example.rdf{e,x}` file, this will add:
- `Example.script.autogen.EXT` for the `Script` content
- `Example.dyncred-script.autogen.EXT` for the `DynamicCredentialScript`
  content (if present)

The `.EXT` above represents the appropriate file extension for
each script interpreter, e.g.: `.py` for Python scripts or
`.ps1` for PowerShell scripts, etc.

A header comment is inserted at the top of each file, noting
that it is auto-generated and manual edits will be lost.